Revert "Change world generation. Refactor command registration. Initial block scanner implementation"
This reverts commit c4b24e06c3d790f752323dc60ce3657f25152257.
