+++ build.gradle.kts
@@ -1,15 +1,15 @@
-import com.github.jengelman.gradle.plugins.shadow.tasks.ShadowJar
-import org.jetbrains.kotlin.gradle.dsl.JvmTarget
-import org.jetbrains.kotlin.gradle.dsl.KotlinVersion
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
+/*
+ * This file was generated by the Gradle 'init' task.
+ */
 
 plugins {
-    java
+    `java-library`
+    `maven-publish`
     kotlin("jvm") version "1.9.0"
-    id("com.github.johnrengelman.shadow") version "8.1.1"
 }
 
 repositories {
+    mavenLocal()
     maven {
         url = uri("https://repo.papermc.io/repository/maven-public/")
     }
@@ -25,15 +25,14 @@ repositories {
     maven {
         url = uri("https://repo.maven.apache.org/maven2/")
     }
-    mavenLocal()
     mavenCentral()
 }
 
 dependencies {
-    implementation("dev.jorel:commandapi-bukkit-shade:9.0.4-SNAPSHOT")
-    implementation("dev.jorel:commandapi-bukkit-kotlin:9.0.4-SNAPSHOT")
-    implementation("org.jetbrains.kotlin:kotlin-stdlib:1.9.0")
-    implementation("com.google.code.gson:gson:2.10.1")
+    api("dev.jorel:commandapi-bukkit-shade:9.0.4-SNAPSHOT")
+    api("dev.jorel:commandapi-bukkit-kotlin:9.0.4-SNAPSHOT")
+    api("org.jetbrains.kotlin:kotlin-stdlib:1.9.0")
+    testImplementation("org.jetbrains.kotlin:kotlin-test:1.9.0")
     compileOnly("io.papermc.paper:paper-api:1.20.1-R0.1-SNAPSHOT")
 }
 
@@ -42,23 +41,20 @@ version = "1.0-SNAPSHOT"
 description = "Infinity"
 java.sourceCompatibility = JavaVersion.VERSION_17
 
-tasks.withType<JavaCompile> {
-    options.encoding = "UTF-8"
-}
-
-tasks.withType<Javadoc> {
-    options.encoding = "UTF-8"
-}
-
-tasks.withType<KotlinCompile> {
-    compilerOptions.languageVersion.set(KotlinVersion.KOTLIN_1_9)
-    compilerOptions.jvmTarget.set(JvmTarget.JVM_17)
-}
-
-tasks.withType<ShadowJar> {
-    dependencies {
-        include(dependency("dev.jorel:commandapi-bukkit-shade:9.0.4-SNAPSHOT"))
-        include(dependency("org.jetbrains.kotlin:kotlin-stdlib:1.9.0"))
+publishing {
+    publications.create<MavenPublication>("maven") {
+        from(components["java"])
     }
-    relocate("dev.jorel.commandapi", "io.github.derechtepilz.commandapi")
+}
+
+tasks.withType<JavaCompile>() {
+    options.encoding = "UTF-8"
+}
+
+tasks.withType<Javadoc>() {
+    options.encoding = "UTF-8"
+}
+
+kotlin {
+    jvmToolchain(17)
 }
